uefi-config: flatten directory structure
Move all the files into the same directory.

This should simplify installation for other users.  I guess that
I will want to separate them again in a better way if I make this
an official tool that gets installed.  But the old structure was
not the way to do it properly anyway.

Signed-off-by: Ryan Harkin <ryan.harkin@linaro.org>