Medium-level redesign of uefi-build.sh
The script has been somewhat redesigned to make it more modular,
and also to add a build summary at the end. Also, the "uefishell"
function is now called only once.

The modularity changes reduce the requirements for adding a new
platform to the script to:
 - adding a <shortname> to the "boards" array
 - creating a <shortname>_LONGNAME variable with a descriptive name
 - creating a <shortname>_build function

The build summary additions prints a log towards the end, specifying
which platforms built successfully, and which failed. The script now
also uses the number of failed builds for its return value.