Introduce single genericised build function
The need for per-platform build functions was making adding new
platforms a bit of a handful. This patch introcudes a build_GENERIC
function which makes use of <platform>_BUILDFLAGS and <platform>_DSC
variables to build the image.

Platforms that do not use standard edk2 'build' (I'm looking at you,
Panda) can override this with a <platform>_BUILDCMD variable.