uefi-backup-branches: add prefix parameter
Allow the use to supply 1 parameter to act as a prefix to the backup branch names.

Before this change, all backup branches were prefixed with "previous-".  Now, "previous-" is only the default prefix, should one not be provided.

Note: the user does not provide the "-" in the prefix name, eg:

  uefi-backup-branches.sh someprefix

Will create backup branches like someprefix-linaro-tracking.

Signed-off-by: Ryan Harkin <ryan.harkin@linaro.org>