uefi-restore-branches: added
Add a new script to restore the branches from the backup version.

Note: this script is broken already! The "branches" list is generated from the list of origina branch names, but if the user deleted one, it won't be in the list.  We need to create a list by stripping off the prefix.

Signed-off-by: Ryan Harkin <ryan.harkin@linaro.org>