Added uefi-vexpress-test.exp
Added an expect script to test UEFI booting on a Versatile Express board in a loop.

Each time, it logs the test as pass or fail and reports the current stats to the user.

It only supports booting an Android image at the moment, but it is trivial to detect an ubuntu or other shell.

Signed-off-by: Ryan Harkin <ryan.harkin@linaro.org>