uefi-new-rc: create tracking branches and handle -rc1
This script would only add -rc2 onwards.  If the monthly tracking branch
or -rc1 did not exist, it would produce the wrong output.

Worth noting is that the internal branch -rc is created before the public
-rc.  This is because the common functions force the next -rc for the
internal tree to be the same as the next -rc for the public tree.  I'm
sure there is something smarter than could be done to handle this, but for
now, it works.

Signed-off-by: Ryan Harkin <ryan.harkin@linaro.org>