uefi-push: push to default remotes or specify on cmdline
Allow the user to push to two different remotes other than the defaults
by adding two parameters to the command line.

It used to only support over-riding the uefi-next remote, but it
occurred to me that if the user was over-riding uefi-next, they were
probably running a test push and therefore should not be pushing to the
uefi-internal remote at the same time.

Signed-off-by: Ryan Harkin <ryan.harkin@linaro.org>