uefi-common: use local config file to over-ride default settings
Previously, a local config file would be sourced instead of the default one.  Now, the default one is sourced and the local one can be used to over-ride any settings from the default config.

Eg. most users will only want to change the UEFI_BASE_DIR variable if they use the same layout as defined in the maintainership wiki.

Signed-off-by: Ryan Harkin <ryan.harkin@linaro.org>