Return True when option not found
Return true when looking up an option that is not found. Returning
failure in this case causes the the build script to exit (on error),
and not finding an option is a 'normal' case, as configs that don't
have to build ATF don't have that option.  We don't want to require
all configs to have every option that exists.

Signed-off-by: Roy Franz <roy.franz@linaro.org>