Use commit id as default FIRMWARE_VER
If a FIRMWARE_VER setting is not provided and we are building from a GIT
repo, provide a default value based on the SHA ID of the commit where we
built from.

If the repo has local uncommited changes, then we mark the version
string with "-dirty".

This is based loosely on the same idea from the linux kernel in:
    scripts/setlocalversion

Signed-off-by: Ryan Harkin <ryan.harkin@linaro.org>