platforms.config: add SCP blobs to Juno EXTRA_FILES
bl30.bin and bl0.bin need to be kept in lock-step.
EXTRA_FILES gives parse-platforms.py a way to extract
the list of all files that need to be installed in
order to update a platform. So add this entry.

This was already done for the platforms.config in LinaroPkg,
actually used for the LEG UEFI releases.

Signed-off-by: Leif Lindholm <leif.lindholm@linaro.org>