Only translate BL30 -> SCP_BL2 when actually specified
This patch ensures that the SCP_BL2 flag is exported only if the
platform configuration has defined the path of the BL30 image. This flag
was always set for all platforms for ARM TF versions greater than 1. If
included in the FVP build, it made the FIP generation fail since SCP_BL2
would point to a directory (no BL30) instead of a file.

Signed-off-by: Achin Gupta <achin.gupta@arm.com>
Signed-off-by: Leif Lindholm <leif.lindholm@linaro.org>