MM: Add SPM_BIN to support prebuilt secure partition image.
This patch adds SPM_BIN to suport prebuilt secure partition image.
For the moment, it assumes that secure partition image is
STANDALONESMM image. The path is relative to
$EDK2_DIR/Build/StandaloneSmmPkg/$BUILD_PROFILE/FV/.

Also, this patch adds commandline "SPM=1" implicitly to build flags,
when SPM_BIN is defined.

SPM_BIN prebuilt image is passed in as a bl32 image to
arm trusted firmware.

Signed-off-by: Supreeth Venkatesh <supreeth.venkatesh@arm.com>
Reviewed-by: Leif Lindholm <leif.lindholm@linaro.org>